Added an account class
